Chat/requirements.txt
amqp==5.2.0
anyio==4.1.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.5
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==40.0.2
daphne==4.0.0
djando==0.2
Django==4.2.1
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-rest-knox==4.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
dynaconf==3.2.5
exceptiongroup==1.2.0
h11==0.14.0
httptools==0.6.1
hvac==2.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
kombu==5.3.4
loguru==0.7.2
msgpack==1.0.5
pillow==10.3.0
prompt-toolkit==3.0.41
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
service-identity==21.1.0
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.28.3
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.10
websockets==12.0
zope.interface==6.0
